pyproject.toml
[tool.poetry]
name = "audit"
version = "0.1.0"
description = "AI Certified Audit"
authors = ["AndreIglesias <ciglesia@student.42.fr>"]
readme = "docs/README.md"

[tool.poetry.dependencies]
python = "^3.10"
requests = "^2.32.3"
flask = "^3.0.3"
jsonify = "^0.5"
numpy = "^2.0.0"
pandas = "^2.2.2"
transformers = "^4.41.2"
gitpython = "^3.1.43"
python-dotenv = "^1.0.1"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
